refactor(html): tree construction "in body"
