+++ html/parser/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "resworb-html-parser"
-version = "0.2.127"
+version = "0.2.128"
 license-file = "../../LICENSE"
 edition = "2021"
